Update exception thrown in ShrubberyCreationForm::execute
It was necessary to change the type of exception thrown when the Bureaucrat's grade is too low to execute the form. Previously, a generic GradeTooLowException was thrown, but the code has been updated to throw a more specific Bureaucrat::GradeTooLowException. This change was made to allow better understanding of errors during execution.